execcmd: manager start task Windows/Linux code decoupling + implementation
