+++ agent/engine/execcmd/manager_start_windows_test.go
@@ -1,4 +1,4 @@
-// +build windows
+// +build windows,unit
 
 // Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
 //
@@ -12,15 +12,10 @@
 // on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 // express or implied. See the License for the specific language governing
 // permissions and limitations under the License.
+
 package execcmd
 
-import apicontainer "github.com/aws/amazon-ecs-agent/agent/api/container"
-
 const (
-	execAgentCmdUser   = "NT AUTHORITY\\SYSTEM"
-	execAgentCmdBinDir = "C:\\Program Files\\Amazon\\SSM"
+	specTestCmd = "C:\\Program Files\\Amazon\\SSM\\amazon-ssm-agent.exe"
+	specUser    = "NT AUTHORITY\\SYSTEM"
 )
-
-func getExecAgentCmdBinDir(ma *apicontainer.ManagedAgent) string {
-	return execAgentCmdBinDir
-}
